RBC-51 | Gopi,Shriram | Handled if Lowest Level Vaule is Not Available
Sometimes Clients might add new AddressHierarchy Level and only new patient who are registered will have this new Level value . To handle older patients who doesn't have this levelValue we are checking if lowerLevel value is not available we are going a step back and checking n-1 lowerLevelValue. This we are doing till the parent level.

Co-Authored-By: Shri Ram <26056117+shriram-k@users.noreply.github.com>